requirements.txt
anthropic>=0.40.0
python-telegram-bot>=20.7
apscheduler>=3.10.4
pyyaml>=6.0.1
python-dotenv>=1.0.0
